refactor: 콘솔 어플리케이션과 스프링 분리
