+++ src/main/java/roomescape/ConsoleRoomescapeApplication.java
@@ -1,8 +1,6 @@
-package roomescape.console;
+package roomescape;
 
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Component;
+import roomescape.console.ConsoleController;
 import roomescape.console.fake.FakeReservationDao;
 import roomescape.console.fake.FakeReservationDb;
 import roomescape.console.fake.FakeReservationTimeDao;
@@ -12,16 +10,14 @@ import roomescape.dao.ReservationTimeDao;
 import roomescape.service.ReservationService;
 import roomescape.service.ReservationTimeService;
 
-@Profile("!test")
-@Component
-public class ConsoleRunner implements CommandLineRunner {
-    @Override
-    public void run(String... args) {
+public class ConsoleRoomescapeApplication {
+
+    public static void main(String[] args) {
         ConsoleController consoleController = initConsoleController();
         consoleController.run();
     }
 
-    private ConsoleController initConsoleController() {
+    private static ConsoleController initConsoleController() {
         FakeReservationDb fakeReservationDb = new FakeReservationDb();
         FakeReservationTimeDb fakeReservationTimeDb = new FakeReservationTimeDb();
         ReservationDao fakeReservationDao
